+++ NEWS.md
@@ -15,6 +15,7 @@
 * Scripting: Fixed missing null check in Tileset.loadFromImage and Tile.setImage
 * Scripting: Initialize tile layer size also when added as part of a group layer (#3291)
 * AutoMapping: Applying rules without "inputnot" layers is now much faster
+* AutoMapping: Optimized calculation of each rule's input/output region
 * AutoMapping: Fixed compatibility with "RegionsInput" / "RegionsOutput" layers
 * terraingenerator: Fixed crash when source terrain doesn't have an image (#3299)
 * macOS: Fixed main window expanding with many open files (#1047)